Rename file from ettu to ettutv
