+++ yt_dlp/extractor/_extractors.py
@@ -527,7 +527,7 @@ from .ellentube import (
 )
 from .elonet import ElonetIE
 from .elpais import ElPaisIE
-from .ettu import EttuTvIE
+from .ettutv import EttuTvIE
 from .embedly import EmbedlyIE
 from .engadget import EngadgetIE
 from .epicon import (
